# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  handlers.PasswordResetEmailRequest:
    properties:
      name:
        type: string
      reset_code:
        type: string
      to:
        type: string
    required:
    - name
    - reset_code
    - to
    type: object
...
  handlers.ResendVerificationRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  services.EmailRequest:
    properties:
      bcc:
        items:
          type: string
        type: array
      body:
        type: string
      cc:
        items:
          type: string
        type: array
      is_html:
        type: boolean
      subject:
        type: string
      template_id:
        type: string
      template_vars:
        additionalProperties: true
        type: object
      to:
        items:
          type: string
        type: array
    required:
    - subject
    - to
    type: object
  services.EmailResponse:
    properties:
      message:
        type: string
      sent_at:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: support@forgecrud.com
    name: API Support
    url: http://www.forgecrud.com/support
  description: Central API documentation - For all ForgeCRUD microservices
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ForgeCRUD API
  version: "1.0"
paths:
  /api/notifications/email/password-reset:
    post:
      consumes:
      - application/json
      description: Send a password reset email with reset code using template
      parameters:
      - description: Password reset email request
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/handlers.PasswordResetEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.EmailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Send password reset email
      tags:
      - email
  /api/notifications/email/resend-verification:
    post:
      consumes:
      - application/json
      description: Resend verification email to user after creating new token
      parameters:
      - description: Resend verification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Resend verification email
      tags:
      - email
  /api/notifications/email/send:
    post:
      consumes:
      - application/json
      description: Send an email through the notification service
      parameters:
      - description: Email request
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/services.EmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.EmailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Send email
      tags:
      - email
  /api/notifications/email/verification:
    post:
      consumes:
      - application/json
      description: Send email verification link to user
      parameters:
      - description: Verification email request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.VerificationEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Send verification email
      tags:
      - email
  /api/notifications/email/welcome:
    post:
      consumes:
      - application/json
      description: Send a welcome email with verification code using template
      parameters:
      - description: Welcome email request
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/handlers.WelcomeEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.EmailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Send welcome/verification email
      tags:
      - email
...
  /documents:
    get:
      consumes:
      - application/json
      description: Retrieve all documents in a specified folder
      parameters:
      - description: Folder ID to list documents from
        in: query
        name: folder_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of documents
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing or invalid folder_id
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get documents in a folder
      tags:
      - documents
    post:
      consumes:
      - multipart/form-data
      description: Upload a new document to a specified folder
      parameters:
      - description: Folder ID where the document will be uploaded
        in: formData
        name: folder_id
        required: true
        type: string
      - description: User ID (for testing purposes)
        in: formData
        name: user_id
        type: string
      - description: Document file to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Document tags
        in: formData
        name: tags
        type: string
      - description: Document description
        in: formData
        name: description
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Document uploaded successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload a new document
      tags:
      - documents
  /documents/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a document and all its versions from storage and database
      parameters:
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Document deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid document ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Document not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a document
      tags:
      - documents
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific document
      parameters:
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Document details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid document ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Document not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get document by ID
      tags:
      - documents
    put:
      consumes:
      - multipart/form-data
      description: Update document tags and description
      parameters:
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Updated document tags
        in: formData
        name: tags
        type: string
      - description: Updated document description
        in: formData
        name: description
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Document updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid document ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Document not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update document metadata
      tags:
      - documents
  /documents/{id}/copy:
    post:
      consumes:
      - application/json
      description: Copy a document and its latest version to a different folder with
        automatic naming
      parameters:
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Target folder information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CopyDocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Document copied successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request data or document ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Document or target folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Copy document to another folder
      tags:
      - documents
  /documents/{id}/download:
    get:
      consumes:
      - application/json
      description: Download the actual file content of a document
      parameters:
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Document file content
          schema:
            type: file
        "400":
          description: Invalid document ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Document not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error or storage unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download document file
      tags:
      - documents
  /documents/{id}/move:
    post:
      consumes:
      - application/json
      description: Move a document and all its versions to a different folder
      parameters:
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Target folder information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.MoveDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Document moved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request data or document ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Document or target folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Move document to another folder
      tags:
      - documents
  /documents/{id}/versions:
    get:
      consumes:
      - application/json
      description: Retrieve all versions of a specific document ordered by version
        number
      parameters:
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of document versions
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Document not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all versions of a document
      tags:
      - documents
    post:
      consumes:
      - multipart/form-data
      description: Upload a new version of an existing document
      parameters:
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Document file to upload
        in: formData
        name: file
        required: true
        type: file
      - description: User ID (for testing purposes)
        in: formData
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Document version uploaded successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Document not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload new document version
      tags:
      - documents
  /documents/{id}/versions/latest:
    get:
      consumes:
      - application/json
      description: Retrieve the most recent version of a specific document
      parameters:
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Latest document version
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Document or version not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get latest version of a document
      tags:
      - documents
  /folders:
    get:
      consumes:
      - application/json
      description: Get all folders with pagination, filtering, sorting and search
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Search term across name and path
        in: query
        name: search
        type: string
      - description: Filter by owner ID
        in: query
        name: filters[owner_id]
        type: string
      - description: Filter by owner type (user, organization)
        in: query
        name: filters[owner_type]
        type: string
      - description: Filter by parent folder ID
        in: query
        name: filters[parent_id]
        type: string
      - description: Sort field (name, path, created_at, updated_at, file_count, total_size)
        in: query
        name: sort[field]
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: sort[order]
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of folders with pagination
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all folders
      tags:
      - folders
    post:
      consumes:
      - application/json
      description: Create a new folder with the provided information
      parameters:
      - description: Folder information
        in: body
        name: folder
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateFolderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created folder
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Folder already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new folder
      tags:
      - folders
...
  /folders/{id}/contents:
    get:
      consumes:
      - application/json
      description: Get all subfolders and documents in a specific folder
      parameters:
      - description: Folder ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Folder contents
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid folder ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get folder contents
      tags:
      - folders
  /folders/{id}/download:
    get:
      consumes:
      - application/json
      description: Download a folder and all its contents as a ZIP archive (recursive)
      parameters:
      - description: Folder ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/zip
      responses:
        "200":
          description: ZIP archive containing folder contents
          schema:
            type: file
        "400":
          description: Invalid folder ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download folder as ZIP
      tags:
      - folders
...
  /ws/notifications/{user_id}:
    get:
      description: Establish WebSocket connection for real-time notifications
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      responses: {}
      summary: WebSocket Connection
      tags:
      - websocket
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...